Initial attempt at go docs
